Extract viewEngine into modules
